requirements.txt
Flask
requests
google-cloud-firestore